use decimal instead of double here, as 1000000000000000000 is past 2^53 - 1 (we'll just cast to double in the end)

+++ src/BizHawk.Client.Common/movie/bk2/Bk2Movie.cs
@@ -98,8 +98,8 @@ namespace BizHawk.Client.Common
 			{
 				if (SystemID == VSystemID.Raw.MAME && Header.TryGetValue("VsyncAttoseconds", out var vsyncAttoStr))
 				{
-					const double attosInSec = 1000000000000000000;
-					return attosInSec / Convert.ToUInt64(vsyncAttoStr);
+					const decimal attosInSec = 1000000000000000000;
+					return (double)(attosInSec / Convert.ToUInt64(vsyncAttoStr));
 				}
 				else
 				{
